4 player functions implemented

+++ src/main/java/com/jasekraft/splendor/mvc/controllers/GameApi.java
@@ -168,7 +168,7 @@ public class GameApi {
     @CrossOrigin(origins = "http://localhost:3000")
     @RequestMapping(value="/games/{gameId}/takenoble/{playerId}", method = RequestMethod.POST)
     public Game takeNoble(@PathVariable("gameId") Long gameId, @PathVariable("playerId") Long playerId, @RequestBody Map<String, Object> body) {
-    	Game game = playerServ.addNoble(gameId, playerId, (Long)body.get("noble_id"));
+    	Game game = playerServ.addNoble(gameId, playerId, Long.valueOf((Integer)body.get("noble_id")));
     	return game;
     }
 }
